project almost completed

+++ Backend/src/controllers/bookController.js
@@ -25,20 +25,38 @@ export const createBook = async (req, res) => {
 
 export const getBook = async (req, res) => {
     try {
-        const { tags, status, page=1, limit=10 } = req.query;
+        const { tags, status, page = 1, limit = 10 } = req.query;
 
         const filter = { user: req.user._id };
         let skip = 0;
-        if(tags) filter.tags = { $in: tags.split(',')};
-        if(status) filter.status = status;
-        
-        if(page && limit){
+        if (status) filter.status = status;
+        if (tags.length) {
+            const tagArray = tags.split(',').map(tag =>
+                new RegExp(`^${tag.trim()}$`, 'i')
+            );
+            filter.tags = { $in: tagArray };
+        }
+        if (page && limit) {
             skip = (+(page) - 1) * +(limit);
         }
 
         const books = await BookModel.find(filter).skip(skip).limit(+limit);
-        
-        res.status(201).json({ message: "Book fetched successfully", data: books });
+        const totalBooks = await BookModel.find({ user: req.user._id }).countDocuments();
+        const wantReadBooks = await BookModel.find({ ...filter, status: "Want to Read" }).countDocuments();
+        const readingBooks = await BookModel.find({ ...filter, status: "Reading" }).countDocuments();
+        const completedBooks = await BookModel.find({ ...filter, status: "Completed" }).countDocuments();
+
+        res.status(201).json({
+            message: "Book fetched successfully",
+            data: books,
+            totalPages: Math.ceil(totalBooks / +limit),
+            bookStatus: {
+                total: totalBooks,
+                wantToRead: wantReadBooks,
+                reading: readingBooks,
+                completed: completedBooks
+            }
+        });
     } catch (error) {
         res.status(500).json({ message: 'Internal Server Error', error: error.message });
     }
@@ -48,8 +66,9 @@ export const updateBook = async (req, res) => {
     try {
         const { title, author, tags, status } = req.body;
         const { bookId } = req.params;
+        const bookStatus = ["Want to Read", "Reading", "Completed"];
 
-        if(!bookId){
+        if (!bookId) {
             return res.status(400).json({ message: "Book ID is required" });
         }
 
@@ -58,10 +77,18 @@ export const updateBook = async (req, res) => {
             return res.status(404).json({ message: "Book not found" });
         }
 
-        if(title.trim()) existBook.title = title;
-        if(author.trim()) existBook.author = author;
-        if(tags.length) existBook.tags = tags;
-        if(status.trim()) existBook.status = status;
+        if(existBook.user.toString() !== req.user._id.toString()){
+            return res.status(403).json({ message: "You are not authorized to update this book" });
+        }
+
+        if (status.trim() && !bookStatus.includes(status.trim())) {
+            return res.status(400).json({ message: 'Provide correct book status' });
+        }
+
+        if (title.trim()) existBook.title = title;
+        if (author.trim()) existBook.author = author;
+        if (tags.length) existBook.tags = tags;
+        if (status.trim()) existBook.status = status;
 
         await existBook.save();
 
@@ -74,13 +101,17 @@ export const updateBook = async (req, res) => {
 export const deleteBook = async (req, res) => {
     try {
         const { bookId } = req.params;
-        if(!bookId){
+        if (!bookId) {
             return res.status(400).json({ message: "Book ID is required" });
         }
 
         const existBook = await BookModel.findById(bookId);
         if (!existBook) {
             return res.status(404).json({ message: "Book not found" });
+        }
+
+        if(existBook.user.toString() !== req.user._id.toString()){
+            return res.status(403).json({ message: "You are not authorized to delete this book" });
         }
 
         await BookModel.findByIdAndDelete(bookId);
